Global flags for device features
The current approach to check for certain device features (like pibridge
functionality) with lots of if-clauses is cumbersome to maintain. It is also
highly likely to forget to add one device to one of the many if-clauses if
a new device is introduced.

Therefore introduce three global flags in `spiControlDev`:

pibridge_rs485_supported: Device supports extension modules with RS485 based
communication (eg. DIO or AIO)

pibridge_ethernet_p2p_supported: Device supports extension modules with point
to point ethernet communication (eg. Gateways)

pibridge_ethernet_supported: Device supports multiple extension modules with
ethernet based communcation (for future use).

Initialize these flags in `piControlInit()` once and use them wherever
appropriate. If a new device is introduced it is sufficient to add it to the
if-clauses in the init function and set the supported flags.

Suggested-by: Lukas Wunner <lukas@wunner.de>
Signed-off-by: Nicolai Buchwitz <n.buchwitz@kunbus.com>