Decrease sniff high period to 8900 us
The IO modules require the sniff pin high period to be within 7500 and 9500 us.

The sniff pins are connected to a GPIO expander on Connect 4 and not to the CM
gpios like other RevPis. Therefore the sniff pins must be written in a sleepable
context with `gpiod_set_value_cansleep` since commit 31ce26694733665022337249aed7304d246a6eaa.

An investigation of recent issues with IO modules which weren't detected after
a restart of the kernel driver, have shown that the period on Connect 4 can be
up to 9505 us. This will force the IO modules into recent state and they cannot
be detected.

Fix this by decreasing the `usleep_range` between high and low to 8500 us.
This should leave enough headroom.

Signed-off-by: Nicolai Buchwitz <n.buchwitz@kunbus.com>